Simplify bytecode version validators
